Move expect_top to lua_util.h
